requirements.txt
Django~=3.2.9
django-filter==21.1
django-widget-tweaks==1.4.9
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==21.1
Pillow==8.4.0
django-allauth==0.46.0